# [testbed] Add batcher perf tests for heavy processing