@swiatekm 2024-12-19
@jmacd @sfc-gh-sili could you have a look? Is this benchmark representative enough to actually help?